+++ tests/unit/migration.spec.ts
@@ -17,107 +17,35 @@ describe("Migration", () => {
     expect(migration).toMatchSnapshot();
   });
 
-  tester.run("v22", migration.rules.v22, {
+  tester.run("v23", migration.rules.v23, {
     valid: [
-      `client.provide("get /v1/test", {id: 10});`,
-      `new Integration({ routing });`,
-      `import { Request } from "./client.ts";`,
-      `createConfig({ cors: true });`,
-      `new Documentation();`,
-      `new EndpointsFactory(new ResultHandler());`,
-      `new EventStreamFactory({});`,
-      `new Client();`,
+      `import { HeaderSecurity } from "express-zod-api";`,
+      `createConfig({ wrongMethodBehavior: 405 });`,
     ],
     invalid: [
       {
-        code: `client.provide("get", "/v1/test", {id: 10});`,
-        output: `client.provide("get /v1/test", {id: 10});`,
+        code: `const security: CustomHeaderSecurity = {};`,
+        output: `const security: HeaderSecurity = {};`,
         errors: [
           {
             messageId: "change",
             data: {
-              subject: "arguments",
-              from: `"get", "/v1/test"`,
-              to: `"get /v1/test"`,
+              subject: "interface",
+              from: "CustomHeaderSecurity",
+              to: "HeaderSecurity",
             },
           },
         ],
       },
       {
-        code: `new Integration({ routing, splitResponse: true });`,
-        output: `new Integration({ routing,  });`,
+        code: `createConfig({});`,
+        output: `createConfig({wrongMethodBehavior: 404,});`,
         errors: [
           {
-            messageId: "remove",
-            data: { subject: "property", name: "splitResponse" },
-          },
-        ],
-      },
-      {
-        code: `import { MethodPath } from "./client.ts";`,
-        output: `import { Request } from "./client.ts";`,
-        errors: [
-          {
-            messageId: "change",
-            data: { subject: "type", from: "MethodPath", to: "Request" },
-          },
-        ],
-      },
-      {
-        code: `createConfig({ tags: { users: "" } });`,
-        output:
-          `createConfig({  });\n` +
-          `// Declaring tag constraints\n` +
-          `declare module "express-zod-api" {\n` +
-          `  interface TagOverrides {\n` +
-          `    "users": unknown,\n` +
-          `  }\n` +
-          `}`,
-        errors: [
-          { messageId: "remove", data: { subject: "property", name: "tags" } },
-        ],
-      },
-      {
-        code: `new Documentation({ config });`,
-        output: `new Documentation({ tags: { /* move from createConfig() argument if any */ }, config });`,
-        errors: [
-          { messageId: "add", data: { subject: "tags", to: "Documentation" } },
-        ],
-      },
-      {
-        code: `new EndpointsFactory({config, resultHandler: new ResultHandler() });`,
-        output: `new EndpointsFactory(new ResultHandler());`,
-        errors: [
-          {
-            messageId: "change",
+            messageId: "add",
             data: {
-              subject: "argument",
-              from: "object",
-              to: "ResultHandler instance",
-            },
-          },
-        ],
-      },
-      {
-        code: `new EventStreamFactory({ config, events: { some } });`,
-        output: `new EventStreamFactory({ some });`,
-        errors: [
-          {
-            messageId: "change",
-            data: { subject: "argument", from: "object", to: "events map" },
-          },
-        ],
-      },
-      {
-        code: `new ExpressZodAPIClient();`,
-        output: `new Client();`,
-        errors: [
-          {
-            messageId: "change",
-            data: {
-              subject: "class",
-              from: "ExpressZodAPIClient",
-              to: "Client",
+              subject: "wrongMethodBehavior property",
+              to: "configuration",
             },
           },
         ],